requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
Django==5.1.4
django-allauth==65.2.0
django-appconf==1.0.6
django-bootstrap5==24.3
django-channels==0.7.0
django-haystack==3.3.0
django-image-cropping==1.7
djangorestframework==3.15.2
easy-thumbnails==2.10
idna==3.10
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Whoosh==2.7.4
